+++ decidim-assemblies/app/commands/decidim/assemblies/admin/update_assembly.rb
@@ -69,6 +69,7 @@ module Decidim
             scopes_enabled: form.scopes_enabled,
             scope: form.scope,
             area: form.area,
+            private_space: form.private_space,
             developer_group: form.developer_group,
             local_area: form.local_area,
             target: form.target,