+++ decidim-assemblies/app/forms/decidim/assemblies/admin/assembly_form.rb
@@ -34,6 +34,7 @@ module Decidim
         attribute :show_statistics, Boolean
         attribute :area_id, Integer
         attribute :participatory_processes_ids, Array[Integer]
+        attribute :private_space, Boolean
 
         validates :slug, presence: true, format: { with: Decidim::Assembly.slug_format }
         validates :title, :subtitle, :description, :short_description, translatable_presence: true