+++ decidim-core/lib/decidim/core.rb
@@ -51,6 +51,7 @@ module Decidim
   autoload :NewsletterEncryptor, "decidim/newsletter_encryptor"
   autoload :QueryExtensions, "decidim/query_extensions"
   autoload :ParticipatorySpaceResourceable, "decidim/participatory_space_resourceable"
+  autoload :HasPrivateUsers, "decidim/has_private_users"
   include ActiveSupport::Configurable
 
   # Loads seeds from all engines.