+++ decidim-participatory_processes/app/commands/decidim/participatory_processes/admin/create_participatory_process.rb
@@ -51,6 +51,7 @@ module Decidim
             promoted: form.promoted,
             scopes_enabled: form.scopes_enabled,
             scope: form.scope,
+            private_space: form.private_space,
             developer_group: form.developer_group,
             local_area: form.local_area,
             target: form.target,
@@ -99,6 +100,16 @@ module Decidim
                    )
 
             Decidim::CreateFollow.new(form, admin).call
+          end
+        end
+
+        def create_participatory_process_users(process)
+          return unless form.private_process
+          form.users.each do |user|
+            ParticipatoryProcessUser.create!(
+              participatory_process: process,
+              user: user
+            )
           end
         end
       end